Tweaks to make labor Items work properly.
