+++ OJewelry/Views/Styles/_CostComponentsPartial.cshtml
@@ -416,7 +416,7 @@ COST:
                 }
                 @Html.Partial("_CostComponentCol1")
 
-                <div class="col-sm-3 col-xs-3">li-LABOR</div>
+                <div class="col-sm-3 col-xs-3">LABOR</div>
                 <div class="col-sm-1 col-xs-1">$/HR</div>
                 <div class="col-sm-1 col-xs-1">$/PC</div>
                 <div class="col-sm-2 col-xs-2">VENDOR</div>
@@ -459,10 +459,10 @@ COST:
                             @if (Model.SVMOp != SVMOperation.Print)
                             {
                                 @*@Html.EditorFor(model => model.LaborItems[i].Name, new { htmlAttributes = new { @class = "col-sm-3 col-xs-3 requiredifnotremoved", @data_val = "true", @data_val_required = "The Labor Name field is required.", @placeholder = "Name" } })*@
-                                @Html.DropDownListFor(model => model.LaborItems[i].laborItemId, new SelectList(Model.drpLaborItems, "Id", "Name", Model.LaborItems[i].laborItemId), "Choose a Labor", new { @class = "col-sm-3 col-xs-3 requiredifnotremoved", @data_val_required = "The Labor Name field is required."  })
-                                @Html.EditorFor(model => model.LaborItems[i].pph, new { htmlAttributes = new { disabled = "disabled", @class = "locked col-sm-1 col-xs-1", onblur = "CalcRowTotal('" + Model.SVMCCType + "', '" + i + "')" } })
-                                @Html.EditorFor(model => model.LaborItems[i].ppp, new { htmlAttributes = new { disabled = "disabled", @class = "locked col-sm-1 col-xs-1", onblur = "CalcRowTotal('" + Model.SVMCCType + "', '" + i + "')" } })
-                                @Html.EditorFor(model => model.LaborItems[i].VendorName, new { htmlAttributes = new { disabled = "disabled", @class = "locked col-sm-2 col-xs-2", onblur = "CalcRowTotal('" + Model.SVMCCType + "', '" + i + "')" } })
+                                @Html.DropDownListFor(model => model.LaborItems[i].laborItemId, new SelectList(Model.drpLaborItems, "Id", "Name", Model.LaborItems[i].laborItemId), "Choose a Labor", new { @class = "col-sm-3 col-xs-3 requiredifnotremoved", onchange="LaborItemsDropdownChanged(" + i + ")", @data_val_required = "The Labor Name field is required."  })
+                                @Html.EditorFor(model => model.LaborItems[i].pph, new { htmlAttributes = new { @class = "nocursor col-sm-1 col-xs-1", onblur = "CalcRowTotal('" + Model.SVMCCType + "', '" + i + "')" } })
+                                @Html.EditorFor(model => model.LaborItems[i].ppp, new { htmlAttributes = new { @class = "nocursor col-sm-1 col-xs-1", onblur = "CalcRowTotal('" + Model.SVMCCType + "', '" + i + "')" } })
+                                @Html.EditorFor(model => model.LaborItems[i].VendorName, new { htmlAttributes = new { @class = "nocursor col-sm-2 col-xs-2", onblur = "CalcRowTotal('" + Model.SVMCCType + "', '" + i + "')" } })
                                 @Html.HiddenFor(model => model.LaborItems[i].Name)
 
                                 <div class="col-sm-1 col-xs-1 "></div>
@@ -523,7 +523,7 @@ COST:
                 }
                 @Html.Partial("_CostComponentCol1")
 
-                <div class="col-sm-3 col-xs-3">l-LABOR</div>
+                <div class="col-sm-3 col-xs-3">LABOR</div>
                 <div class="col-sm-3 col-xs-3">Desc</div>
                 <div class="col-sm-1 col-xs-1">$/HR</div>
                 <div class="col-sm-1 col-xs-1">$/PC</div>
